rename to gen-ai-attributes, and fix test
