+++ dev-packages/node-integration-tests/suites/tracing/openai/instrument-with-options.mjs
@@ -1,9 +1,6 @@
 import * as Sentry from '@sentry/node';
 import { loggingTransport } from '@sentry-internal/node-integration-tests';
 
-// Set environment variable to trigger the fourth test
-process.env.TEST_OPTIONS = '1';
-
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
   release: '1.0',
