+++ packages/core/src/utils/gen-ai-attributes.ts
